[IMP] mail: move follower button menu into FollowerList component
